refactor(Modal): remove default export and extract common modal components
refs: CDS-26
